Refactor: Today page creation
- I created the new page where I move all the tasks creation functions
- I remove all the selectors and functions that are not used
